+++ src/config.js
@@ -1,0 +1,6 @@
+export const teamColors = {
+    BLUE: "linear-gradient(45deg, #1d2e6e, #1d4e8e)",
+    RED: "linear-gradient(45deg, #700, #b00)",
+    NEUTRAL: "linear-gradient(45deg, #444, #777)",
+    DEATH: "linear-gradient(45deg, #000, #111)",
+}
