+++ src/socket.js
@@ -1,0 +1,5 @@
+import io from "socket.io-client"
+
+const socket = io()
+
+export default socket
